Check usernames and displayNames against profane/misleading word list (#103)
This PR checks usernames and displayNames against a list of
profane/misleading words and throws a validation error if they're found.
I check for exact matches (e.g. "sex") and also for matches with extra
numbers in it (e.g. "sex1"). It'd be pretty easy to get around this
system with various characters that look like other characters, or
phonetic spellings, but I expect this will be good enough, and we can
improve it if that's ever needed. (One can be overzealous in applying
profanity filters; this is known as the [Scunthorpe
Problem](https://en.wikipedia.org/wiki/Scunthorpe_problem).)

[Associated ClickUp task](https://app.clickup.com/t/86866f10z).