Implement option to disable message requests (#153)
This PR adds a user option to disable message requests. It's unchecked by
default.

When a user has checked it, the Message button doesn't show on their
profile (except for admins, in which case there's some text telling the
admin that they've disabled it but admins can still message them). We
also check on the server, in messageCallbacks.

[ClickUp task](https://app.clickup.com/t/8685a3vfu).