Change default comment order to a setting (#168)
This PR changes the default comment order to use a new
post.defaultCommentOrder setting.

[ClickUp task](https://app.clickup.com/t/8686r8qd3).

This is intended for merging upstream (which I'll arrange after it's
finalized and approved here). EAForum will want their setting to be
"postCommentsMagic", and AlignmentForum will want "afPostCommentsTop".
LessWrong and WakingUp will use the default, "postCommentsTop".