Add dev and staging favicons (#169)
The staging favicons are served from Cloudinary, but I also add them to
git just for reference. The dev ones are actually served from the web
server, though.

The new favicons are the color of ClickUp Ready For QA (in staging), and
the color of In Progress (for dev).

Here's [the ClickUp task](https://app.clickup.com/t/8686q0563).