Update WU generic error message to include a mailto link (#175)
This PR updates the standard error message. The message includes an HTML
anchor tag, so we want to use `dangerouslySetInnerHTML` to make the link
get interpreted. We wouldn't want to do this with any error message,
because I'm not confident no-one will ever include some user-generated
content in an error message. So we check that the error message matches
the generic error message exactly, and if so, it allows HTML interpretation.

[ClickUp task](https://app.clickup.com/t/8686tdywe).